requirements.txt
streamlit
streamlit-audiorecorder
openai
python-dotenv
pytube
pydub